[filegen] move plansdk.FlakePlan and plansdk.FlakeInput
