+++ internal/impl/devbox.go
@@ -33,7 +33,6 @@ import (
 	"go.jetpack.io/devbox/internal/impl/devopt"
 	"go.jetpack.io/devbox/internal/lock"
 	"go.jetpack.io/devbox/internal/nix"
-	"go.jetpack.io/devbox/internal/planner/plansdk"
 	"go.jetpack.io/devbox/internal/plugin"
 	"go.jetpack.io/devbox/internal/redact"
 	"go.jetpack.io/devbox/internal/searcher"
@@ -136,52 +135,6 @@ func (d *Devbox) ConfigHash() (string, error) {
 
 func (d *Devbox) NixPkgsCommitHash() string {
 	return d.cfg.NixPkgsCommitHash()
-}
-
-func (d *Devbox) FlakePlan(ctx context.Context) (*plansdk.FlakePlan, error) {
-	ctx, task := trace.NewTask(ctx, "devboxFlakePlan")
-	defer task.End()
-
-	// Create plugin directories first because inputs might depend on them
-	for _, pkg := range d.PackagesAsInputs() {
-		if err := d.pluginManager.Create(pkg); err != nil {
-			return nil, err
-		}
-	}
-
-	for _, included := range d.cfg.Include {
-		// This is a slightly weird place to put this, but since includes can't be
-		// added via command and we need them to be added before we call
-		// plugin manager.Include
-		if err := d.lockfile.Add(included); err != nil {
-			return nil, err
-		}
-		if err := d.pluginManager.Include(included); err != nil {
-			return nil, err
-		}
-	}
-
-	shellPlan := &plansdk.FlakePlan{}
-	var err error
-	shellPlan.FlakeInputs, err = d.flakeInputs(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	nixpkgsInfo := plansdk.GetNixpkgsInfo(d.cfg.NixPkgsCommitHash())
-
-	// This is an optimization. Try to reuse the nixpkgs info from the flake
-	// inputs to avoid introducing a new one.
-	for _, input := range shellPlan.FlakeInputs {
-		if input.IsNixpkgs() {
-			nixpkgsInfo = plansdk.GetNixpkgsInfo(input.HashFromNixPkgsURL())
-			break
-		}
-	}
-
-	shellPlan.NixpkgsInfo = nixpkgsInfo
-
-	return shellPlan, nil
 }
 
 func (d *Devbox) Generate(ctx context.Context) error {
@@ -1174,4 +1127,12 @@ func ExportifySystemPathWithoutWrappers() string {
 	}
 
 	return exportify(envs)
+}
+
+func (d *Devbox) PluginManager() *plugin.Manager {
+	return d.pluginManager
+}
+
+func (d *Devbox) Lockfile() *lock.File {
+	return d.lockfile
 }
